Use trace to handle scalar matrix multiplication results
Sympy has been very limited and frustrating. When multiplying a 1xn and nx1 matrices, the product is treated as 1x1 matrix instead of scalar leading to bugs (if you multiply the scalar with another matrix, the result h=end up having one single row).

Accessing the element [0] of the matrix multiplication result to use it as a scalar ends up with sympy forcibly expanding all the terms even when explictly told not to.

For now, using trace() operator to have sympy handle the inner product as a scalar while not expanding the terms